+++ src/utils/componentCategories.ts
@@ -1,42 +1,38 @@
 // NYPL Design System Component Categories
 
 const categories = {
-  content: {
-    title: "Content Display",
+  basicContent: {
+    title: "Components/Basic Elements",
+    components: ["Card", "CardEdition", "Hero", "Promo", "Sponsor", "Table"],
+  },
+  contentDisplay: {
+    title: "Components/Content Display",
     components: [
       "AccessibilityDisplayHelper",
-      "Accordion",
       "AddressDisplayHelper",
-      "Card",
-      "ContentSwitcher",
       "DateDisplayHelper",
-      "DefinitionList",
-      "Heading",
-      "Hero",
+      "HelperErrorText",
       "HoursDisplayHelper",
-      "List",
-      "Modal",
       "OverflowMenu",
-      "Promo",
+      "StatusBadge",
     ],
   },
   deprecated: {
-    title: "Deprecated",
+    title: "Components/Deprecated",
     components: [],
   },
   feedback: {
-    title: "Feedback",
+    title: "Components/Feedback",
     components: [
       "Alert",
       "InlineLoading",
       "Notification",
       "ProgressIndicator",
       "SkeletonLoader",
-      "Tooltip",
     ],
   },
   forms: {
-    title: "Form Elements",
+    title: "Components/Form Elements",
     components: [
       "Button",
       "Checkbox",
@@ -59,24 +55,53 @@ const categories = {
     ],
   },
   layout: {
-    title: "Page Layout",
-    components: ["BasicPageSection", "Footer", "Grid", "Header", "Template"],
+    title: "Components/Page Layout",
+    components: [
+      "BasicPageSection",
+      "ContentSwitcher",
+      "Footer",
+      "Grid",
+      "Header",
+      "HorizontalRule",
+      "Section",
+      "Template",
+    ],
+  },
+  overlayContentSwitcher: {
+    title: "Components/Overlays & Switchers",
+    components: ["Accordion", "Modal", "Tabs", "Tooltip"],
   },
   media: {
-    title: "Media",
-    components: ["AudioPlayer", "Icon", "Image", "Logo", "Slideshow", "Video"],
+    title: "Components/Media & Icons",
+    components: [
+      "AudioPlayer",
+      "Icon",
+      "Image",
+      "Logo",
+      "Slideshow",
+      "VideoPlayer",
+    ],
   },
   navigation: {
-    title: "Navigation",
+    title: "Components/Navigation",
     components: ["Breadcrumbs", "Link", "Menu", "Pagination", "Subnavigation"],
   },
+  typography: {
+    title: "Components/Typography & Styles",
+    components: ["Heading", "List"],
+  },
+  devguide: {
+    title: "Documentation/Development Guide",
+    components: ["Autosuggest"],
+  },
   styleguide: {
-    title: "Style Guide",
+    title: "Documentation/Style Guide",
     components: [
       "Bidirectionality",
       "Breakpoints",
       "Buttons",
       "Colors",
+      "Forms",
       "Iconography",
       "Spacing",
       "Template Bidirectionality",