+++ CHANGELOG.md
@@ -8,6 +8,30 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 0.24.0
+
+### Breaking Changes
+
+- Renames original `Card` components to be `CardEdition`.
+
+### Adds
+
+- Adds the `DatePicker` component.
+- Adds new `Card` component.
+- Adds `spacing` prop to `Form` component to allow for spacing variations.
+- Adds Style Guide `Forms`
+
+### Changes
+
+- Exports `VideoPlayerAspectRatios`.
+- Adds additional props for the `DatePicker` component based on its TAD Addendum section.
+- Adds Tugboat QA configuration for PR preview deployments to replace Netlify.
+- Updates the `DatePicker`, `TextInput`, `Select`, and `Label` components to pass an optional `showOptReqLabel` prop to conditionally render "Required"/"Optional" in the label text.
+
+### Fixes
+
+- Fixes Accordion FAQ Storybook example rendering bug.
+
 ## 0.23.4
 
 ### Adds
@@ -132,7 +156,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Removes `--ui-success` CSS variable and adds `--ui-success-primary` and `--ui-success-secondary`
 - Removes `checkboxId` prop from `Checkbox` component and employs `id` in its place.
 - Removes `defaultChecked` prop from `Checkbox` component.
-- Removed `labelOptions` porp from `Checkbox` component and employs `labelText` in its place.
+- Removed `labelOptions` prop from `Checkbox` component and employs `labelText` in its place.
 
 ### Changes